+++ src/yolo_annotator_desktop/__init__.py
@@ -1,3 +1,3 @@
 """YOLO Annotator Desktop."""
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"